envelope: move op functions
Move the Envelope-specific functions into libenvelope.
